move scoped trace helpers into Tracer
